+++ tests/Configs/ArrayConfigTest.php
@@ -6,9 +6,11 @@ use Heyday\CacheInclude\Configs\ArrayConfig;
 
 class ArrayConfigTest extends \PHPUnit_Framework_TestCase
 {
+    protected $testData;
+
     protected function setUp()
     {
-        $this->data = array(
+        $this->testData = array(
             'One',
             'Two'
         );
@@ -16,15 +18,14 @@ class ArrayConfigTest extends \PHPUnit_Framework_TestCase
 
     protected function tearDown()
     {
-        $this->data = null;
+        $this->testData = null;
     }
 
     public function testIteration()
     {
-        $config = new ArrayConfig($this->data);
-        $i = 0;
+        $config = new ArrayConfig($this->testData);
         foreach ($config as $key => $val) {
-            $this->assertEquals($this->data[$key], $val);
+            $this->assertEquals($this->testData[$key], $val);
         }
     }
     /**
@@ -33,7 +34,7 @@ class ArrayConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testSet()
     {
-        $config = new ArrayConfig($this->data);
+        $config = new ArrayConfig($this->testData);
         $config[2] = 'Three';
     }
     /**
@@ -42,20 +43,20 @@ class ArrayConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testUnset()
     {
-        $config = new ArrayConfig($this->data);
+        $config = new ArrayConfig($this->testData);
         unset($config[1]);
     }
 
     public function testGet()
     {
-        $config = new ArrayConfig($this->data);
+        $config = new ArrayConfig($this->testData);
         $this->assertEquals('One', $config[0]);
         $this->assertEquals('Two', $config[1]);
     }
 
     public function testExists()
     {
-        $config = new ArrayConfig($this->data);
+        $config = new ArrayConfig($this->testData);
         $this->assertTrue(isset($config[0]));
         $this->assertFalse(isset($config[2]));
     }